Remove classes of com.theta360.ptp.** from signature  of com.theta360.theta

+++ lib/src/main/java/com/theta360/ptp/PtpEventListener.java
@@ -17,7 +17,7 @@ public interface PtpEventListener {
 
     void onStoreRemoved(UINT32 storageID);
 
-    void onDevicePropChanged(UINT32 devicePropCode);
+    void onDevicePropChanged(UINT16 devicePropCode);
 
     void onObjectInfoChanged(UINT32 objectHandle);
